🚚 move enum to util file
